* added cuda internal api header: this is the result of some reverse engineering of the os x cuda driver (346 branch) I did and currently contains the internal texture reference, context, sampler "enum"/type and sampler pool structs; this will be used to set certain sampler state that is currently not exposed by the cuda driver api; note that the struct content is very much specific to the current os x driver version and not compatible to earlier versions or the linux driver yet (also, thanks for providing debug symbols nvidia ;))  * cuda_api: added some os x specific functions/symbols (these are only exported on os x, so no linux/windows equivalents yet)  * cuda texture objects are now only stored in a uint32_t (instead of a uint64_t), since the cuda api will only track/return a 32-bit value anyways (internally it's handled as a 64-bit value, but the upper part (the sampler object) is never returned), might do the same for surface objects (have not checked yet if it's the same)  * llvm_compute / backend devices / device_info: added DEPTH_COMPARE and GATHER capabilities (signals if depth compare and gather are supported in h/w or if s/w emulation is required)  * opencl_buffer: always block when reading/writing the buffer  * host-compute/host_device: updated image capabilities

+++ compute/cuda/cuda_kernel.hpp
@@ -187,8 +187,8 @@ protected:
 		// set texture+sampler objects
 		const auto& textures = cu_img->get_cuda_textures();
 		for(const auto& texture : textures) {
-			memcpy(data, &texture, sizeof(uint64_t));
-			data += sizeof(uint64_t);
+			memcpy(data, &texture, sizeof(uint32_t));
+			data += sizeof(uint32_t);
 		}
 		
 		// set surface object
